Split into two serveless functions, and work with database instead of body params
